Add logic to handle move to next screen.
