+++ app/src/main/AndroidManifest.xml
@@ -13,6 +13,9 @@
             android:name=".views.onboarding.OnboardingActivity"
             android:exported="false"></activity>
         <activity
+            android:name=".views.login.LoginActivity"
+            android:exported="false"></activity>
+        <activity
             android:name=".SplashActivity"
             android:exported="true">
             <intent-filter>
