+++ app/src/main/java/com/qibla/qiblacompass/prayertime/finddirection/BaseActivity.kt
@@ -1,0 +1,22 @@
+package com.qibla.qiblacompass.prayertime.finddirection
+
+import android.os.Bundle
+import android.view.Window
+import android.view.WindowManager
+import androidx.appcompat.app.AppCompatActivity
+
+abstract class BaseActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        makeFullScreen()
+        super.onCreate(savedInstanceState)
+    }
+
+    private fun makeFullScreen() {
+        requestWindowFeature(Window.FEATURE_NO_TITLE)
+        window.setFlags(
+            WindowManager.LayoutParams.FLAG_FULLSCREEN,
+            WindowManager.LayoutParams.FLAG_FULLSCREEN
+        )
+        supportActionBar?.hide()
+    }
+}
