+++ app/src/main/java/com/qibla/qiblacompass/prayertime/finddirection/views/login/LoginActivity.kt
@@ -2,9 +2,10 @@ package com.qibla.qiblacompass.prayertime.finddirection.views.login
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import com.qibla.qiblacompass.prayertime.finddirection.BaseActivity
 import com.qibla.qiblacompass.prayertime.finddirection.R
 
-class LoginActivity : AppCompatActivity() {
+class LoginActivity : BaseActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_login)
